requirements.txt
flask
jinja2
python-dotenv
ytmusicapi

gunicorn

